Move integration test to a separate file
